requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.20.0
scikit-learn>=1.3.0
google-generativeai>=0.8.0
